requirements.txt
absl-py==2.1.0
alembic==1.16.4
amqp==5.3.1
asgiref==3.8.1
astunparse==1.6.3
bcrypt==4.3.0
beautifulsoup4==4.12.2
bidict==0.23.1
billiard==4.2.1
blinker==1.6.2
celery==5.3.4
certifi==2023.5.7
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comtypes==1.1.14
contourpy==1.1.0
cryptography==41.0.7
cssselect==1.2.0
cycler==0.11.0
deepface==0.0.95
Django==5.2.3
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
fire==0.7.1
Flask==2.3.3
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
flatbuffers==24.12.23
fonttools==4.40.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
googletrans==3.0.0
gpt3==0.0.1
greenlet==3.2.3
grpcio==1.68.1
gTTS==2.3.2
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
h5py==3.12.1
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
jaraco.context==4.3.0
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.5.2
keras==3.12.0
kiwisolver==1.4.4
kombu==5.5.4
libclang==18.1.1
lxml==5.3.0
lxml_html_clean==0.2.2
lz4==4.4.4
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
matplotlib==3.8.3
mdurl==0.1.2
ml_dtypes==0.5.3
mlxtend==0.23.4
more-itertools==9.1.0
MouseInfo==0.1.3
mpmath==1.3.0
mtcnn==1.0.0
namex==0.0.8
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
opencv-python==4.7.0.72
opt_einsum==3.4.0
optree==0.13.1
packaging==23.1
pandas==2.1.0
Pillow==10.0.1
playsound==1.3.0
plyer==2.1.0
prompt_toolkit==3.0.51
protobuf==5.29.2
PuLP==2.8.0
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycparser==2.22
pygame==2.4.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyngrok==7.4.1
pynput==1.7.6
pyparsing==3.1.0
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.12.2
python-socketio==5.9.0
PythonTurtle==0.3.2
pyttsx3==2.90
pytweening==1.0.7
pytz==2023.3.post1
pywhatkit==5.4
pywin32==305
PyYAML==6.0
redis==5.0.1
regex==2023.8.8
requests==2.31.0
requests-file==2.1.0
retina-face==0.0.17
rfc3986==1.5.0
rich==13.9.4
scikit-learn==1.5.2
scipy==1.15.2
seaborn==0.13.2
sgmllib3k==1.0.0
simple-websocket==1.1.0
six==1.16.0
sketchpy==0.1.0
sniffio==1.3.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
speedtest==0.0.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
svg.path==6.2
svgpathtools==1.6.0
svgwrite==1.4.3
sympy==1.12
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
textblob==0.18.0.post0
tf_keras==2.20.1
threadpoolctl==3.5.0
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.65.0
turtle==0.0.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.0.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.3.7
whitenoise==6.11.0
wikipedia==1.4.0
wolframalpha==5.0.0
wrapt==1.17.0
wsproto==1.2.0
xmltodict==0.13.0
